+++ packages/sdk/src/sdk-blockchains/tezos/mint.test.ts
@@ -71,9 +71,17 @@ describe("mint test", () => {
 			blockchain: Blockchain.TEZOS,
 			name: "1",
 			description: "2",
-			image: "ipfs://ipfs/QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG5",
-			animationUrl: "ipfs://ipfs/QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG6",
-			externalUrl: "ipfs://ipfs/QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG7",
+			image: {
+				url: "ipfs://ipfs/QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG1",
+				mimeType: "image/jpeg",
+				fileName: "image",
+			},
+			animation: {
+				url: "ipfs://ipfs/QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG2",
+				mimeType: "image/mp4",
+				fileName: "video",
+			},
+			external: "https://rarible.com",
 			attributes: [{
 				key: "eyes",
 				value: "1",
@@ -82,10 +90,20 @@ describe("mint test", () => {
 
 		expect(response.name).toBe("1")
 		expect(response.description).toBe("2")
-		expect(response.artifactUri).toBe("ipfs://ipfs/QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG5")
-		expect(response.displayUri).toBe("ipfs://ipfs/QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG5")
-		expect(response.thumbnailUri).toBe("ipfs://ipfs/QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG6")
-		expect(response.externalUri).toBe("ipfs://ipfs/QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG7")
+		expect(response.artifactUri).toBe("ipfs://QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG2")
+		expect(response.displayUri).toBe("ipfs://QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG1")
+		expect(response.formats).toStrictEqual([
+			{
+				uri: "ipfs://QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG1",
+				mimeType: "image/jpeg",
+				fileName: "image",
+			},
+			{
+				uri: "ipfs://QmfVqzkQcKR1vCNqcZkeVVy94684hyLki7QcVzd9rmjuG2",
+				mimeType: "image/mp4",
+				fileName: "video",
+			},
+		])
 		expect(response.attributes[0].name).toBe("eyes")
 		expect(response.attributes[0].value).toBe("1")
 	})