+++ packages/sdk/src/types/nft/mint/preprocess-meta.ts
@@ -17,20 +17,35 @@ export type TokenMetadataAttribute = {
 export type CommonTokenMetadata = {
 	name: string
 	description: string | undefined
-	image: string | undefined
-	animationUrl: string | undefined
-	externalUrl: string | undefined
+	image: CommonTokenContent | undefined
+	animation: CommonTokenContent | undefined
+	external: string | undefined
 	attributes: TokenMetadataAttribute[]
 	royalties?: UnionPart[]
+}
+
+export type CommonTokenContent = {
+	url: string
+	mimeType: string
+	hash?: string
+	fileSize?: number
+	fileName?: string
+	duration?: string
+	dataRate?: {
+		value: number
+		unit: string
+	}
+	dimensions?: {
+		value: string
+		unit: string
+	}
 }
 
 export type CommonTokenMetadataResponse = {
 	name: string
 	description: string | undefined
 	image: string | undefined
-	// eslint-disable-next-line camelcase
-	animation_url: string | undefined
-	// eslint-disable-next-line camelcase
-	external_url: string | undefined
+	"animation_url": string | undefined
+	"external_url": string | undefined
 	attributes: TokenMetadataAttribute[]
 }