+++ packages/sdk/CHANGELOG.md
@@ -3,6 +3,90 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.43](https://github.com/rarible/sdk/compare/v0.7.42...v0.7.43) (2021-12-17)
+
+
+### Bug Fixes
+
+* restriction msg ([#71](https://github.com/rarible/sdk/issues/71)) ([6247f0f](https://github.com/rarible/sdk/commit/6247f0fe443bf304cce33cafde06cbb4daa0e08f))
+
+
+### Features
+
+* deploy method update ([#72](https://github.com/rarible/sdk/issues/72)) ([5f7a413](https://github.com/rarible/sdk/commit/5f7a41357c1163e7cac47d3339a66a3dba828e0e))
+
+
+
+
+
+## [0.7.42](https://github.com/rarible/sdk/compare/v0.7.41...v0.7.42) (2021-12-16)
+
+
+### Bug Fixes
+
+* update flow-sdk, fix evolution transaction ([d3535d3](https://github.com/rarible/sdk/commit/d3535d3ec567a31988739da62cb742430e931f78))
+
+
+
+
+
+## [0.7.41](https://github.com/rarible/sdk/compare/v0.7.40...v0.7.41) (2021-12-16)
+
+
+### Bug Fixes
+
+* preprocessing of meta ([193c803](https://github.com/rarible/sdk/commit/193c8038e00d57479abbb79b7424114285d35066))
+
+
+
+
+
+## [0.7.40](https://github.com/rarible/sdk/compare/v0.7.39...v0.7.40) (2021-12-16)
+
+
+### Features
+
+* update preprocess meta ([#67](https://github.com/rarible/sdk/issues/67)) ([2c96dca](https://github.com/rarible/sdk/commit/2c96dca8b1deb7d42b3fa6d0660b31fa3c77f0e0))
+
+
+
+
+
+## [0.7.39](https://github.com/rarible/sdk/compare/v0.7.38...v0.7.39) (2021-12-16)
+
+
+### Bug Fixes
+
+* update flow-sdk ([93d4c1f](https://github.com/rarible/sdk/commit/93d4c1f9df1625b95606ebeed435edd2f3f851e1))
+* update flow-sdk version (return legacy order trasactions) ([09e7e01](https://github.com/rarible/sdk/commit/09e7e012273c98d9152fc3463b6057ec6f6b5b60))
+
+
+
+
+
+## [0.7.38](https://github.com/rarible/sdk/compare/v0.7.37...v0.7.38) (2021-12-15)
+
+
+### Bug Fixes
+
+* better tezos metadata (tests) ([b75f89d](https://github.com/rarible/sdk/commit/b75f89d933de2584553c993051a710fa2155cfd3))
+* better tezos metadata + tests ([621f82c](https://github.com/rarible/sdk/commit/621f82cdf158414e7ab1bd2e40bd88bc25205c1d))
+
+
+
+
+
+## [0.7.37](https://github.com/rarible/sdk/compare/v0.7.36...v0.7.37) (2021-12-15)
+
+
+### Bug Fixes
+
+* better tezos metadata ([c30b205](https://github.com/rarible/sdk/commit/c30b2051fdaf7d518c51220077717a35a9aed8a6))
+
+
+
+
+
 ## [0.7.36](https://github.com/rarible/sdk/compare/v0.7.35...v0.7.36) (2021-12-15)
 
 **Note:** Version bump only for package @rarible/sdk